Remove the code related to starcache from the dir block_cache.
