+++ be/src/agent/task_worker_pool.cpp
@@ -48,8 +48,8 @@
 #include "agent/report_task.h"
 #include "agent/resource_group_usage_recorder.h"
 #include "agent/task_signatures_manager.h"
-#include "cache/block_cache/datacache_utils.h"
-#include "cache/block_cache/local_cache.h"
+#include "cache/datacache_utils.h"
+#include "cache/local_cache.h"
 #include "common/status.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/workgroup/work_group.h"
