+++ be/src/bench/object_cache_bench.cpp
@@ -17,10 +17,10 @@
 #include <cstdlib>
 #include <random>
 
-#include "cache/block_cache/cache_options.h"
-#include "cache/block_cache/starcache_wrapper.h"
+#include "cache/cache_options.h"
 #include "cache/object_cache/lrucache_module.h"
 #include "cache/object_cache/starcache_module.h"
+#include "cache/starcache_wrapper.h"
 #include "common/config.h"
 #include "runtime/current_thread.h"
 #include "runtime/exec_env.h"
