+++ be/src/cache/block_cache/block_cache.h
@@ -16,9 +16,9 @@
 
 #include <atomic>
 
-#include "cache/block_cache/disk_space_monitor.h"
-#include "cache/block_cache/local_cache.h"
-#include "cache/block_cache/remote_cache.h"
+#include "cache/disk_space_monitor.h"
+#include "cache/local_cache.h"
+#include "cache/remote_cache.h"
 #include "common/status.h"
 
 namespace starrocks {
