+++ be/src/cache/datacache_utils.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "cache/block_cache/datacache_utils.h"
+#include "cache/datacache_utils.h"
 
 #include <fmt/format.h>
 #include <sys/stat.h>
