+++ be/src/cache/datacache_utils.h
@@ -14,7 +14,7 @@
 
 #pragma once
 
-#include "cache/block_cache/local_cache.h"
+#include "cache/local_cache.h"
 #include "gen_cpp/DataCache_types.h"
 
 namespace starrocks {
