+++ be/src/cache/remote_cache.h
@@ -14,54 +14,35 @@
 
 #pragma once
 
-#include "cache/block_cache/cache_options.h"
 #include "cache/block_cache/io_buffer.h"
+#include "cache/cache_options.h"
 #include "common/status.h"
 
 namespace starrocks {
 
-enum class DataCacheEngineType { STARCACHE };
-
-class LocalCache {
+class RemoteCache {
 public:
-    virtual ~LocalCache() = default;
+    virtual ~RemoteCache() = default;
 
-    // Init KV cache
+    // Init remote cache
     virtual Status init(const CacheOptions& options) = 0;
-    virtual bool is_initialized() const = 0;
 
-    // Write data to cache
+    // Write data to remote cache
     virtual Status write(const std::string& key, const IOBuffer& buffer, WriteCacheOptions* options) = 0;
 
-    // Read data from cache, it returns the data size if successful; otherwise the error status
+    // Read data from remote cache, it returns the data size if successful; otherwise the error status
     // will be returned.
     virtual Status read(const std::string& key, size_t off, size_t size, IOBuffer* buffer,
                         ReadCacheOptions* options) = 0;
 
-    virtual bool exist(const std::string& key) const = 0;
-
     // Remove data from cache.
     virtual Status remove(const std::string& key) = 0;
 
-    // Update the datacache memory quota.
-    virtual Status update_mem_quota(size_t quota_bytes, bool flush_to_disk) = 0;
+    virtual void record_read_remote(size_t size, int64_t lateny_us) = 0;
 
-    // Update the datacache disk space information, such as disk quota or disk path.
-    virtual Status update_disk_spaces(const std::vector<DirSpace>& spaces) = 0;
-
-    virtual const DataCacheMetrics cache_metrics(int level) const = 0;
-
-    virtual void record_read_remote(size_t size, int64_t latency_us) = 0;
-
-    virtual void record_read_cache(size_t size, int64_t latency_us) = 0;
+    virtual void record_read_cache(size_t size, int64_t lateny_us) = 0;
 
     virtual Status shutdown() = 0;
-
-    virtual DataCacheEngineType engine_type() = 0;
-
-    virtual bool available() const = 0;
-    virtual bool mem_cache_available() const = 0;
-    virtual void disk_spaces(std::vector<DirSpace>* spaces) const = 0;
 };
 
 } // namespace starrocks
