+++ be/src/exec/schema_scanner/schema_be_datacache_metrics_scanner.cpp
@@ -15,7 +15,7 @@
 #include "exec/schema_scanner/schema_be_datacache_metrics_scanner.h"
 
 #include "agent/master_info.h"
-#include "cache/block_cache/local_cache.h"
+#include "cache/local_cache.h"
 #include "column/datum.h"
 #include "gutil/strings/substitute.h"
 #include "runtime/exec_env.h"
