+++ be/src/formats/parquet/column_reader.h
@@ -22,7 +22,7 @@
 #include <string>
 #include <vector>
 
-#include "cache/block_cache/cache_options.h"
+#include "cache/cache_options.h"
 #include "column/column.h"
 #include "column/vectorized_fwd.h"
 #include "common/object_pool.h"
