+++ be/src/formats/parquet/file_reader.cpp
@@ -24,7 +24,7 @@
 #include <utility>
 #include <vector>
 
-#include "cache/block_cache/local_cache.h"
+#include "cache/local_cache.h"
 #include "column/column.h"
 #include "column/column_helper.h"
 #include "column/const_column.h"
