+++ be/test/CMakeLists.txt
@@ -524,9 +524,9 @@ set(EXEC_FILES
         ./storage/lake/replication_txn_manager_test.cpp
         ./storage/lake/persistent_index_sstable_test.cpp
         ./storage/lake/write_combined_txn_log_test.cpp
-        ./cache/block_cache/datacache_utils_test.cpp
+        ./cache/datacache_utils_test.cpp
         ./cache/block_cache/block_cache_hit_rate_counter_test.cpp
-        ./cache/block_cache/peer_cache_test.cpp
+        ./cache/peer_cache_test.cpp
         ./cache/object_cache/lrucache_module_test.cpp
         ./cache/object_cache/object_cache_test.cpp
         ./util/thrift_rpc_helper_test.cpp
@@ -535,8 +535,8 @@ set(EXEC_FILES
         )
 
 if ("${WITH_STARCACHE}" STREQUAL "ON")
+    list(APPEND EXEC_FILES ./cache/disk_space_monitor_test.cpp)
     list(APPEND EXEC_FILES ./cache/block_cache/block_cache_test.cpp)
-    list(APPEND EXEC_FILES ./cache/block_cache/disk_space_monitor_test.cpp)
     list(APPEND EXEC_FILES ./cache/object_cache/starcache_module_test.cpp)
     list(APPEND EXEC_FILES ./cache/object_cache/object_cache_test.cpp)
     list(APPEND EXEC_FILES ./io/cache_input_stream_test.cpp)
