+++ be/test/cache/block_cache/block_cache_test.cpp
@@ -19,8 +19,8 @@
 #include <cstring>
 #include <filesystem>
 
-#include "cache/block_cache/datacache_utils.h"
 #include "cache/block_cache/test_cache_utils.h"
+#include "cache/datacache_utils.h"
 #include "common/logging.h"
 #include "common/statusor.h"
 #include "fs/fs_util.h"
