+++ be/test/cache/object_cache/starcache_module_test.cpp
@@ -16,8 +16,8 @@
 
 #include <gtest/gtest.h>
 
-#include "cache/block_cache/starcache_wrapper.h"
 #include "cache/block_cache/test_cache_utils.h"
+#include "cache/starcache_wrapper.h"
 #include "fs/fs_util.h"
 #include "testutil/assert.h"
