+++ be/test/io/cache_input_stream_test.cpp
@@ -17,8 +17,8 @@
 #include <gtest/gtest.h>
 
 #include "cache/block_cache/block_cache.h"
-#include "cache/block_cache/starcache_wrapper.h"
 #include "cache/block_cache/test_cache_utils.h"
+#include "cache/starcache_wrapper.h"
 #include "fs/fs_util.h"
 #include "runtime/exec_env.h"
 #include "testutil/assert.h"
